update default parameters; restructure wake turbulence and combination
