+++ floris/simulation/wake.py
@@ -1,13 +1,14 @@
-# Copyright 2019 NREL
+# Copyright 2020 NREL
 
-# Licensed under the Apache License, Version 2.0 (the "License"); you may not use
-# this file except in compliance with the License. You may obtain a copy of the
-# License at http://www.apache.org/licenses/LICENSE-2.0
+# Licensed under the Apache License, Version 2.0 (the "License"); you may not
+# use this file except in compliance with the License. You may obtain a copy of
+# the License at http://www.apache.org/licenses/LICENSE-2.0
 
-# Unless required by applicable law or agreed to in writing, software distributed
-# under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-# CONDITIONS OF ANY KIND, either express or implied. See the License for the
-# specific language governing permissions and limitations under the License.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations under
+# the License.
 
 from .wake_velocity.base_velocity_deficit import VelocityDeficit
 from .wake_velocity.curl import Curl as CurlDeficit
@@ -16,7 +17,7 @@ from .wake_velocity.gaussianModels.gauss_merge import MergeGauss as MergeGaussDe
 from .wake_velocity.gaussianModels.gauss import Gauss as GaussDeficit
 from .wake_velocity.jensen import Jensen
 from .wake_velocity.multizone import MultiZone
-from .wake_velocity.gaussianModels.ishihara import Ishihara
+from .wake_velocity.gaussianModels.ishihara_qian import IshiharaQian as IshiharaQianDeficit
 from .wake_velocity.gaussianModels.blondel import Blondel as BlondelDeficit
 
 from .wake_deflection.base_velocity_deflection import VelocityDeflection
@@ -24,8 +25,15 @@ from .wake_deflection.jimenez import Jimenez
 from .wake_deflection.gauss import Gauss as GaussDeflection
 from .wake_deflection.curl import Curl as CurlDeflection
 
-from . import wake_turbulence
-from . import wake_combination
+from .wake_turbulence.base_wake_turbulence import WakeTurbulence
+from .wake_turbulence.crespo_hernandez import CrespoHernandez as CrespoHernandezTurbulence
+from .wake_turbulence.ishihara_qian import IshiharaQian as IshiharaQianTurbulence
+from .wake_turbulence.direct import Direct as DirectTurbulence
+
+from .wake_combination.base_wake_combination import WakeCombination
+from .wake_combination.fls import FLS
+from .wake_combination.sosfs import SOSFS
+from .wake_combination.max import MAX
 
 
 class Wake():
@@ -66,16 +74,17 @@ class Wake():
             "gauss": GaussDeficit,
             "gauss_merge": MergeGaussDeficit,
             "gauss_legacy": LegacyGaussDeficit,
-            "ishihara": Ishihara,
+            "ishihara_qian": IshiharaQianDeficit,
             "curl": CurlDeficit,
             "blondel": BlondelDeficit
         }
         self.velocity_model = properties["velocity_model"]
 
         self._turbulence_models = {
-            "gauss": wake_turbulence.Gauss,
-            "ishihara": wake_turbulence.Ishihara,
-            "None": wake_turbulence.WakeTurbulence
+            "crespo_hernandez": CrespoHernandezTurbulence,
+            "ishihara_qian": IshiharaQianTurbulence,
+            "direct": DirectTurbulence,
+            "None": WakeTurbulence
         }
         self.turbulence_model = properties["turbulence_model"]
 
@@ -87,9 +96,9 @@ class Wake():
         self.deflection_model = properties["deflection_model"]
 
         self._combination_models = {
-            "fls": wake_combination.FLS,
-            "sosfs": wake_combination.SOSFS,
-            "max": wake_combination.MAX
+            "fls": FLS,
+            "sosfs": SOSFS,
+            "max": MAX
         }
         self.combination_model = properties["combination_model"]
