+++ floris/simulation/wake_combination/base_wake_combination.py
@@ -1,0 +1,22 @@
+# Copyright 2020 NREL
+
+# Licensed under the Apache License, Version 2.0 (the "License"); you may not
+# use this file except in compliance with the License. You may obtain a copy of
+# the License at http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations under
+# the License.
+
+class WakeCombination():
+    """
+    The WakeCombination class provides methods for combining the base flow field with the velocity deficits from the wake models.
+
+    Returns:
+        WakeCombination: An instantiated WakeCombination object.
+    """
+
+    def __init__(self):
+        self.model_string = None
