+++ floris/simulation/wake_deflection/curl.py
@@ -10,6 +10,7 @@
 # License for the specific language governing permissions and limitations under
 # the License.
 
+from ...utilities import setup_logger
 from .base_velocity_deflection import VelocityDeflection
 import numpy as np
 
@@ -39,6 +40,7 @@ class Curl(VelocityDeflection):
                     }
         """
         super().__init__(parameter_dictionary)
+        self.logger = setup_logger(name=__name__)
         self.model_string = "curl"
 
     def function(self, x_locations, y_locations, z_locations, turbine, coord,
