+++ floris/simulation/wake_velocity/gauss.py
@@ -10,7 +10,7 @@
 # License for the specific language governing permissions and limitations under
 # the License.
 
-from ...utilities import cosd, sind, tand
+from ...utilities import cosd, sind, tand, setup_logger
 from .base_velocity_deficit import VelocityDeficit
 import numpy as np
 
@@ -106,6 +106,7 @@ class Gauss(VelocityDeficit):
 
     def __init__(self, parameter_dictionary):
         super().__init__(parameter_dictionary)
+        self.logger = setup_logger(name=__name__)
         self.model_string = "gauss"
         model_dictionary = self._get_model_dict()
