+++ floris/simulation/wake_velocity/gaussianModels/gauss.py
@@ -11,6 +11,7 @@
 # the License.
 
 from ....utilities import cosd, sind, tand
+from ....utilities import setup_logger
 from ..base_velocity_deficit import VelocityDeficit
 import numpy as np
 
@@ -106,6 +107,7 @@ class Gauss(VelocityDeficit):
 
     def __init__(self, parameter_dictionary):
         super().__init__(parameter_dictionary)
+        self.logger = setup_logger(name=__name__)
         self.model_string = "gauss"
         model_dictionary = self._get_model_dict(__class__.default_parameters)
