+++ floris/simulation/wake_velocity/jensen.py
@@ -10,6 +10,7 @@
 # License for the specific language governing permissions and limitations under
 # the License.
 
+from ...utilities import setup_logger
 from .base_velocity_deficit import VelocityDeficit
 import numpy as np
 
@@ -58,6 +59,7 @@ class Jensen(VelocityDeficit):
 
     def __init__(self, parameter_dictionary):
         super().__init__(parameter_dictionary)
+        self.logger = setup_logger(name=__name__)
         self.model_string = "jensen"
         model_dictionary = self._get_model_dict(__class__.default_parameters)
         self.we = float(model_dictionary["we"])
